+++ src/graphical/Drawers/GraphDrawer.cpp
@@ -18,10 +18,10 @@ void GraphDrawer::putLayer(int layer)
   {
     had_update = updateOneMarker(layer);
 
-    for(unsigned long int i = 0; i < branchs_nodes.size(); i++)
+    for(uint32_t i = 0; i < branchs_nodes.size(); i++)
     {
       bool previous_done = true;
-      for(unsigned long int pre = 0; pre < branchs_nodes[i]->prev.size(); pre++)
+      for(uint32_t pre = 0; pre < branchs_nodes[i]->prev.size(); pre++)
         if(branchs_nodes[i]->prev[pre]->marker == false)
           previous_done = false;
 
@@ -38,15 +38,15 @@ void GraphDrawer::putLayer(int layer)
 
 bool GraphDrawer::updateOneMarker(int layer)
 {
-  for(unsigned long int i = 0; i < roots_nodes.size(); i++)
+  for(uint32_t i = 0; i < roots_nodes.size(); i++)
     if((roots_nodes[i]->layer == layer) && (roots_nodes[i]->marker == false))
     {
       roots_nodes[i]->marker = true;
       return true;
     }
 
-  for(unsigned long int layer_i = 0; layer_i < layer_nodes.size(); layer_i++)
-    for(unsigned long int i = 0; i < layer_nodes[layer_i].size(); i++)
+  for(uint32_t layer_i = 0; layer_i < layer_nodes.size(); layer_i++)
+    for(uint32_t i = 0; i < layer_nodes[layer_i].size(); i++)
       if((layer_nodes[layer_i][i]->layer == layer) && (layer_nodes[layer_i][i]->marker == false))
       {
         layer_nodes[layer_i][i]->marker = true;
@@ -60,7 +60,7 @@ bool GraphDrawer::testEnd()
 {
   bool end = true;
 
-  for(unsigned long int i = 0; i < branchs_nodes.size(); i++)
+  for(uint32_t i = 0; i < branchs_nodes.size(); i++)
     if(branchs_nodes[i]->marker == false)
       end = false;
 
@@ -69,10 +69,10 @@ bool GraphDrawer::testEnd()
 
 void GraphDrawer::draw(std::string file_name)
 {
-  unsigned long int height = (layer_nodes.size())*(MARKER_HEIGHT + MIN_HEIGHT_SPACE) + 1;
+  uint32_t height = (layer_nodes.size())*(MARKER_HEIGHT + MIN_HEIGHT_SPACE) + 1;
 
-  unsigned long int width = roots_nodes.size();
-  for(unsigned long int i  = 0; i < layer_nodes.size(); i++)
+  uint32_t width = roots_nodes.size();
+  for(uint32_t i  = 0; i < layer_nodes.size(); i++)
       if(layer_nodes[i].size() > width)
         width = layer_nodes[i].size();
   width = width*(MARKER_WIDTH + MIN_WIDTH_SPACE) + 1;
@@ -80,11 +80,11 @@ void GraphDrawer::draw(std::string file_name)
   image = cvCreateImage(cvSize(width, height), IPL_DEPTH_8U, 3);
   cvSet(image, cvScalar(255,255,255));
 
-  for(unsigned long int i = 0; i < roots_nodes.size(); i++)
+  for(uint32_t i = 0; i < roots_nodes.size(); i++)
     setRect(0, layer_nodes.size() + 1, roots_nodes.size(), roots_nodes[i]);
 
-  for(unsigned long int layer = 0; layer < layer_nodes.size(); layer++)
-    for(unsigned long int i = 0; i < layer_nodes[layer].size(); i++)
+  for(uint32_t layer = 0; layer < layer_nodes.size(); layer++)
+    for(uint32_t i = 0; i < layer_nodes[layer].size(); i++)
     setRect(layer+1, layer_nodes.size() + 1, layer_nodes[layer].size(), layer_nodes[layer][i]);
 
   link();
@@ -120,9 +120,9 @@ void GraphDrawer::setRect(int layer, int nb_layer, int nb_index, node_t* node)
 
 void GraphDrawer::link()
 {
-  for(unsigned long int layer = 0; layer < layer_nodes.size(); layer++)
-    for(unsigned long int i = 0; i < layer_nodes[layer].size(); i++)
-      for(unsigned long int pre = 0; pre < layer_nodes[layer][i]->prev.size(); pre++)
+  for(uint32_t layer = 0; layer < layer_nodes.size(); layer++)
+    for(uint32_t i = 0; i < layer_nodes[layer].size(); i++)
+      for(uint32_t pre = 0; pre < layer_nodes[layer][i]->prev.size(); pre++)
       {
         cvLine(image, cvPoint(layer_nodes[layer][i]->rect.x_middle_top(),
                               layer_nodes[layer][i]->rect.y_middle_top()),
@@ -134,7 +134,7 @@ void GraphDrawer::link()
 
 bool GraphDrawer::exist(std::string value)
 {
-  for(unsigned long int i = 0; i < branchs_nodes.size(); i++)
+  for(uint32_t i = 0; i < branchs_nodes.size(); i++)
     if(branchs_nodes[i]->value == value)
       return true;
